+++ src/main/java/com/devbliss/gpullr/domain/PullRequest.java
@@ -17,6 +17,8 @@ import javax.validation.constraints.NotNull;
 @Entity
 public class PullRequest {
 
+  private static final String SEPARATOR = " / ";
+  
   public enum State {
     OPEN, CLOSED
   }
@@ -88,5 +90,10 @@ public class PullRequest {
     }
 
     return true;
+  }
+
+  @Override
+  public String toString() {
+    return "[id=" + repo.id + SEPARATOR + repo.name + SEPARATOR + number + "]";
   }
 }