+++ src/main/java/com/devbliss/gpullr/util/http/GithubHttpClientImplNoop.java
@@ -1,6 +1,6 @@
 package com.devbliss.gpullr.util.http;
 
-import com.devbliss.gpullr.service.github.GetGithubEventsRequest;
+import com.devbliss.gpullr.service.github.AbstractGithubRequest;
 
 /**
  * Implementation that does no real calls. To be used as mock, for tests etc.
@@ -11,7 +11,7 @@ import com.devbliss.gpullr.service.github.GetGithubEventsRequest;
 public class GithubHttpClientImplNoop implements GithubHttpClient {
 
   @Override
-  public GithubHttpResponse execute(GetGithubEventsRequest request) {
+  public GithubHttpResponse execute(AbstractGithubRequest request) {
     return null;
   }
 }